refactor: 마이크 테스트 모달 로직 분리
